+++ CHANGELOG.md
@@ -3,6 +3,7 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 
 ## Unreleased
 - Added model function for exporting program as CSV
+- Added management command for purging test programs
 
 ## 2.2.0
 - Release for school review